Function iota_adapter_latest::adapter::new_native_extensions

pub fn new_native_extensions<'r>(
    child_resolver: &'r dyn ChildObjectResolver,
    input_objects: BTreeMap<ObjectID, InputObject>,
    is_metered: bool,
    protocol_config: &'r ProtocolConfig,
    metrics: Arc<LimitsMetrics>,
    current_epoch_id: EpochId,
) -> NativeContextExtensions<'r>

Creates a new set of NativeContextExtensions for the Move VM, configuring extensions such as ObjectRuntime and NativesCostTable. These extensions manage object resolution, input objects, metering, protocol configuration, and metrics tracking. They are available and mainly used in native function implementations via NativeContext instance.