iota_adapter_latest::execution_engine

Function authenticate_then_execute_transaction_to_effects

pub fn authenticate_then_execute_transaction_to_effects<Mode: ExecutionMode>(
    store: &dyn BackingStore,
    protocol_config: &ProtocolConfig,
    metrics: Arc<LimitsMetrics>,
    enable_expensive_checks: bool,
    certificate_deny_set: &HashSet<TransactionDigest>,
    epoch_id: &EpochId,
    epoch_timestamp_ms: u64,
    gas_data: GasData,
    gas_status: IotaGasStatus,
    authenticators: Vec<(MoveAuthenticator, AuthenticatorFunctionRefForExecution, CheckedInputObjects)>,
    authenticator_and_transaction_input_objects: CheckedInputObjects,
    transaction_kind: TransactionKind,
    transaction_signer: IotaAddress,
    transaction_digest: TransactionDigest,
    transaction_data_bytes: Vec<u8>,
    trace_builder_opt: &mut Option<MoveTraceBuilder>,
    move_vm: &Arc<MoveVM>,
) -> (InnerTemporaryStore, IotaGasStatus, TransactionEffects, Result<Mode::ExecutionResults, ExecutionError>)

This function produces transaction effects for a transaction that requires the Move authentication. It creates a temporary store, gas charger, and transaction context for the authentication execution and then reuses these for the normal transaction execution. Running the Move authentication can have two outcomes:

If it fails, then it charges gas for the failed execution of the authentication and produces transaction effects with the appropriate error status.

Else, if the authentication is successful, it continues with the normal transaction execution. It combines the input objects from both the authentication and transaction.