iota_adapter_latest::execution_engine

Function authenticate_transaction_inner

pub fn authenticate_transaction_inner(
    temporary_store: &mut TemporaryStore<'_>,
    protocol_config: &ProtocolConfig,
    metrics: Arc<LimitsMetrics>,
    gas_charger: &mut GasCharger,
    authenticator: MoveAuthenticator,
    authenticator_function_ref: AuthenticatorFunctionRefV1,
    authenticator_input_objects: &InputObjects,
    transaction_kind: TransactionKind,
    transaction_digest: TransactionDigest,
    tx_data_bytes: Vec<u8>,
    tx_ctx: Rc<RefCell<TxContext>>,
    trace_builder_opt: &mut Option<MoveTraceBuilder>,
    move_vm: &Arc<MoveVM>,
) -> Result<<Authentication as ExecutionMode>::ExecutionResults, ExecutionError>

MoveAuthenticator PTB with a single move call for execution, then executes it through an inner execution method. The MoveAuthenticator provides the inputs to use for the authentication function found in AuthenticatorFunctionRef, that is retrieved from an account. If the execution fails, it returns an execution error; otherwise it returns an empty value.