iota_adapter_latest::execution_engine

Function construct_advance_epoch_pt_impl

pub fn construct_advance_epoch_pt_impl(
    builder: ProgrammableTransactionBuilder,
    params: &AdvanceEpochParams,
    call_arg_vec: Vec<CallArg>,
) -> Result<ProgrammableTransaction, ExecutionError>

Constructs a ProgrammableTransaction to advance the epoch. It creates storage charges and computation charges by invoking mint_epoch_rewards_in_pt, advances the epoch by setting up the necessary arguments, such as epoch number, protocol version, storage rebate, and slashing rate, and executing the advance_epoch function within the IOTA system. Then, it destroys the storage rebates to complete the transaction.