iota_adapter_latest::execution_engine

Function execute_genesis_state_update

pub fn execute_genesis_state_update(
    store: &dyn BackingStore,
    protocol_config: &ProtocolConfig,
    metrics: Arc<LimitsMetrics>,
    move_vm: &Arc<MoveVM>,
    tx_context: &mut TxContext,
    input_objects: CheckedInputObjects,
    pt: ProgrammableTransaction,
) -> Result<InnerTemporaryStore, ExecutionError>

Function dedicated to the execution of a GenesisTransaction. The function creates an InnerTemporaryStore, processes the input objects, and executes the transaction in unmetered mode using the Genesis execution mode. It returns an inner temporary store that contains the objects found into the input GenesisTransaction by adding the data for previous_transaction and storage_rebate fields.