Module iota_adapter_latest::adapter

Functions

missing_unwrapped_msg

Returns an error message for a missing unwrapped object.

new_move_vm

Creates a new instance of MoveVM with the specified native functions and protocol configuration. The VM is configured using a VMConfig that sets limits for vector length, value depth, and other runtime options based on the provided ProtocolConfig. If gas profiling is enabled, the function configures the profiler with the provided path.

new_native_extensions

Creates a new set of NativeContextExtensions for the Move VM, configuring extensions such as ObjectRuntime and NativesCostTable. These extensions manage object resolution, input objects, metering, protocol configuration, and metrics tracking. They are available and mainly used in native function implementations via NativeContext instance.

run_metered_move_bytecode_verifier

Run the bytecode verifier with a meter limit

substitute_package_id

Given a list of modules and an object_id, mutate each module’s self ID (which must be 0x0) to be object_id.