iota_adapter_latest

Module execution_engine

Functions

construct_advance_epoch_pt_impl

Constructs a ProgrammableTransaction to advance the epoch. It creates storage charges and computation charges by invoking mint_epoch_rewards_in_pt, advances the epoch by setting up the necessary arguments, such as epoch number, protocol version, storage rebate, and slashing rate, and executing the advance_epoch function within the IOTA system. Then, it destroys the storage rebates to complete the transaction.

construct_advance_epoch_pt_v1

construct_advance_epoch_pt_v2

construct_advance_epoch_pt_v3

execute_genesis_state_update

Function dedicated to the execution of a GenesisTransaction. The function creates an InnerTemporaryStore, processes the input objects, and executes the transaction in unmetered mode using the Genesis execution mode. It returns an inner temporary store that contains the objects found into the input GenesisTransaction by adding the data for previous_transaction and storage_rebate fields.

execute_transaction_to_effects

The main entry point to the adapter’s transaction execution. It prepares a transaction for execution, then executes it through an inner execution method and finally produces an instance of transaction effects. It also returns the inner temporary store, which contains the objects resulting from the transaction execution, the gas status instance, which tracks the gas usage, and the execution result. The function handles transaction execution based on the provided TransactionKind. It checks for any expensive operations, manages shared object references, and ensures transaction dependencies are met. The returned objects are not committed to the store until the resulting effects are applied by the caller.