Function iota_adapter_latest::execution_engine::execute_transaction_to_effects

pub fn execute_transaction_to_effects<Mode: ExecutionMode>(
    store: &dyn BackingStore,
    input_objects: CheckedInputObjects,
    gas_coins: Vec<ObjectRef>,
    gas_status: IotaGasStatus,
    transaction_kind: TransactionKind,
    transaction_signer: IotaAddress,
    transaction_digest: TransactionDigest,
    move_vm: &Arc<MoveVM>,
    epoch_id: &EpochId,
    epoch_timestamp_ms: u64,
    protocol_config: &ProtocolConfig,
    metrics: Arc<LimitsMetrics>,
    enable_expensive_checks: bool,
    certificate_deny_set: &HashSet<TransactionDigest>,
) -> (InnerTemporaryStore, IotaGasStatus, TransactionEffects, Result<Mode::ExecutionResults, ExecutionError>)

The main entry point to the adapter’s transaction execution. It prepares a transaction for execution, then executes it through an inner execution method and finally produces an instance of transaction effects. It also returns the inner temporary store, which contains the objects resulting from the transction execution, the gas status instance, which tracks the gas usage, and the execution result. The function handles transaction execution based on the provided TransactionKind. It checks for any expensive operations, manages shared object references, and ensures transaction dependencies are met. The returned objects are not committed to the store until the resulting effects are applied by the caller.